requirements.txt
amqp==5.1.1
asgiref==3.6.0
auth==0.5.3
beautifulsoup4==4.12.2
billiard==3.6.4.0
blinker==1.5
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==39.0.1
Django==4.1.6
django-cors-headers==4.2.0
django-jazzmin==2.6.0
django-jsonfield==1.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
drf-writable-nested==0.7.0
drf-yasg==1.21.7
eventlet==0.33.3
falcon==3.1.1
flower==1.2.0
googletrans==4.0.0rc1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
humanize==4.6.0
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
joblib==1.3.1
kombu==5.2.4
MarkupSafe==2.1.2
mongoengine==0.26.0
nltk==3.8.1
packaging==23.1
pexels-api==1.0.1
Pillow==9.4.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
pypexels==1.0.0rc1
python-decouple==3.8
pytz==2022.7.1
PyYAML==6.0.1
regex==2023.6.3
requests==2.28.2
rfc3986==1.5.0
ruwordnet==0.0.6
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.49
sqlparse==0.4.3
tornado==6.2
tqdm==4.65.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
